requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask

# mysql libraries
python-dotenv
mysql-connector-python 
pymysql

#-e .
